\name{densEstBayesVignette}
\alias{densEstBayesVignette}
\title{Display the package's vignette.}
\description{The vignette of the densEstBayes package is displayed using the default PDF file browser. It provides a detailed detailed description of use of the package for Bayesian density estimation.}
\usage{
densEstBayesVignette()
}
\author{Matt P. Wand \email{matt.wand@uts.edu.au}}
\examples{
\donttest{densEstBayesVignette()}
}


